integration testing: rename reactions.rs to timeline.rs
