+++ testing/matrix-sdk-integration-testing/src/tests.rs
@@ -2,9 +2,9 @@ mod auth;
 mod e2ee;
 mod invitations;
 mod nse;
-mod reactions;
 mod redaction;
 mod repeated_join;
 mod room;
 mod room_directory_search;
 mod sliding_sync;
+mod timeline;
